+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [0.1.3] – 2025-04-25
 
 ### Changed
@@ -63,3 +65,4 @@ First *eriplots* release!
 [0.1.1]: https://github.com/ElderResearch/eriplots-python/releases/tag/0.1.1
 [0.1.2]: https://github.com/ElderResearch/eriplots-python/releases/tag/0.1.2
 [0.1.3]: https://github.com/ElderResearch/eriplots-python/releases/tag/0.1.3
+[Unreleased]: https://github.com/ElderResearch/eriplots-python/compare/0.1.3...develop