+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- (DEV) Added a release checklist.
+
 ## [0.1.3] – 2025-04-25
 
 ### Changed